Mini code refactoring
